requirements.txt
# Django 2.1 requires an update to mozilla-django-oidc.
# See: https://github.com/mozilla/mozilla-django-oidc/issues/266
Django==2.2
PyOpenSSL==19.0.0
contextlib2==0.5.5
djangorestframework==3.9.2
dj-database-url==0.5.0
dockerflow==2018.4.0
factory_boy==2.11.1
gunicorn==19.9.0
psycopg2==2.8.1
python-decouple==3.1
raven==6.10.0
requests==2.21.0
simplejson==3.16.0
six==1.12.0

# For linting and testing.
configparser==3.7.4
flake8==3.7.7
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
coverage==4.5.3
